Fixed sync error at start of game if a player changes their options just when the game host has started the game
The options request takes priority over the "start game" message, so while the game had already
started for the game host, the game host still handled the options request and changed the options
for non-host players before sending them the "start game" message, resulting in the in-game player
options being different for the game host and the non-host players. This naturally led to a desync
right at the start of the game.