Resolve the generic hostname last.
If the generic host is resolved as one of the explicit hosts, the server
will only be tried with the generic 6667 port instead of the
server-specific port values, due to the duplicate-checker code filtering
out the explicit host since it's DNS record returns after the generic
one. Placing the generic DNS record as last ensures that all specific
hosts will be resolved beforehand and allows connections in cases
where port 6667 is unreachable due to a network issue.