Make CampaignID default to -1, with an optional override
Different parsing of duplicate entries in Battle.ini between the client and the game has led to the game crashing
as the game sees duplicate entries as one entry, while the client sees them as separate entries.
Separate entries are often used by modders to add separators to the mission list.

The game is not known to really use the campaign ID for much (only the "do you need GDI/Nod/Firestorm CD" check is known),
so setting it to -1 by default should do little harm.
In case it does, there is now the possibility to manually override the campaign ID.